+++ src/panel/panelIndex.html
@@ -2,11 +2,15 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
+    <meta
+      http-equiv="Content-Security-Policy"
+      content="default-src 'none'; script-src 'nonce-${nonce}'; style-src 'nonce-${nonce}' ${webview.cspSource};"
+    />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Panel Webview</title>
   </head>
   <body>
     <div id="root"></div>
-    <script src="panel.bundle.js"></script>
+    <script nonce="${nonce}" src="panel.js"></script>
   </body>
 </html>
