+++ src/sidebar/sidebarIndex.html
@@ -15,49 +15,3 @@
     <script nonce="${nonce}" src="../../dist/sidebar.js"></script>
   </body>
 </html>
-
-<!-- 
-    <link href="/styles/sidebar.css" rel="stylesheet" />
-
-
-
-
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta
-      http-equiv="Content-Security-Policy"
-      content="default-src 'none'; script-src 'nonce-${nonce}'; style-src ${webview.cspSource};"
-    />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>SusCode</title>
-  </head>
-  <body class="body">
-    <div id="root">
-      <button class="get-extensions-button">Get Extensions</button>
-      <div class="extensions-list"></div>
-    </div>
-    <script>
-      window.vscode = acquireVsCodeApi();
-    </script>
-    <script>
-      function getNonce() {
-        let text = '';
-        const possible =
-          'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
-        for (let i = 0; i < 32; i++) {
-          text += possible.charAt(Math.floor(Math.random() * possible.length));
-        }
-        return text;
-      }
-      const nonce = getNonce();
-    </script>
-    <script nonce="${nonce}" src=""></script>
-  </body>
-</html> -->
-
-<!-- <meta
-http-equiv="Content-Security-Policy"
-content="default-src 'none'; script-src 'nonce-${nonce}'; style-src ${webview.cspSource};"
-/> -->
