+++ src/vscode.d.ts
@@ -1,0 +1,5 @@
+declare function acquireVsCodeApi(): {
+  postMessage: (msg: any) => void;
+  setState: <T>(newState: T) => void;
+  getState: <T>() => T | undefined;
+};
